refactor(scm-messaging-integrations-ui): extract useScmMessagingIntegrationsQuery to own file
